requirements.txt
streamlit
pandas
requests
langchain_openai
pydantic
python-dotenv
